+++ packages/react/src/components/F0Button/F0Button.tsx
@@ -13,6 +13,7 @@ const privateProps = [
   "compact",
   "noTitle",
   "noAutoTooltip",
+  "style",
 ] as const
 
 export type F0ButtonProps = Omit<