+++ packages/react/src/components/F0ButtonDropdown/types.ts
@@ -30,6 +30,11 @@ export type ButtonDropdownItem<T = string> = {
   description?: string
 }
 
+export type ButtonDropdownGroup<T = string> = {
+  label?: string
+  items: ButtonDropdownItem<T>[]
+}
+
 export type F0ButtonDropdownProps<T = string> = {
   /**
    * The size of the button.
@@ -39,7 +44,10 @@ export type F0ButtonDropdownProps<T = string> = {
   /**
    * The items to display in the dropdown.
    */
-  items: ButtonDropdownItem<T>[]
+  items:
+    | ButtonDropdownItem<T>[]
+    | ButtonDropdownGroup<T>[]
+    | ButtonDropdownGroup<T>
   /**
    * The variant of the button.
    * @default "default"