+++ packages/react/src/experimental/OneActionBar/index.tsx
@@ -2,15 +2,12 @@ import { F0AvatarAlert } from "@/components/avatars/F0AvatarAlert"
 import { F0Button } from "@/components/F0Button"
 import { F0ButtonDropdown } from "@/components/F0ButtonDropdown"
 import { IconType } from "@/components/F0Icon"
-import {
-  Dropdown,
-  DropdownItem,
-  MobileDropdown,
-} from "@/experimental/Navigation/Dropdown"
+import { Dropdown, MobileDropdown } from "@/experimental/Navigation/Dropdown"
 import { useI18n } from "@/lib/providers/i18n"
 import { cn } from "@/lib/utils"
 import NumberFlow from "@number-flow/react"
 import { AnimatePresence, motion } from "motion/react"
+import { Fragment, useCallback, useMemo } from "react"
 
 type ActionType = {
   label: string
@@ -21,7 +18,42 @@ type ActionType = {
   description?: string
 }
 
-type ActionBarItem = ActionType | { type: "separator" }
+export type ActionBarGroup = {
+  label?: string
+  items: ActionBarItem[]
+}
+
+export type ActionBarItem = ActionType
+
+function isActionGroup(
+  item: ActionBarItem | ActionBarGroup
+): item is ActionBarGroup {
+  return "items" in item
+}
+
+/**
+ * Normalize the items to an array of DropdownButtonGroup
+ */
+const normalizeItems = (
+  items: ActionType[] | ActionBarGroup[] | ActionBarGroup
+): ActionBarGroup[] => {
+  if (Array.isArray(items)) {
+    if (items.every((item) => isActionGroup(item))) {
+      // ActionBarGroup[]
+      return items
+    } else {
+      // ActionType[]
+      return [
+        {
+          items: items,
+        },
+      ]
+    }
+  } else {
+    // ActionBarGroup
+    return [items]
+  }
+}
 
 interface OneActionBarProps {
   /**
@@ -32,7 +64,7 @@ interface OneActionBarProps {
   /**
    * The primary action
    */
-  primaryActions?: ActionBarItem[]
+  primaryActions?: ActionBarItem[] | ActionBarGroup[] | ActionBarGroup
 
   /**
    * The secondary actions
@@ -56,10 +88,6 @@ interface OneActionBarProps {
   warningMessage?: string
 }
 
-function isActionType(item: ActionBarItem): item is ActionType {
-  return "label" in item
-}
-
 const Alert = ({ message }: { message: string }) => {
   return (
     <div className="flex w-full flex-row items-center gap-2 rounded-md bg-f1-background-warning p-2">
@@ -71,11 +99,11 @@ const Alert = ({ message }: { message: string }) => {
 
 export const OneActionBar = ({
   isOpen,
-  primaryActions = [],
   secondaryActions = [],
   selectedNumber = undefined,
   onUnselect,
   warningMessage,
+  ...props
 }: OneActionBarProps) => {
   const i18n = useI18n()
 
@@ -84,17 +112,56 @@ export const OneActionBar = ({
       ? i18n.status.selected.singular
       : i18n.status.selected.plural
 
-  const visibleSecondaryActions = secondaryActions
-    .slice(0, 2)
-    .filter(isActionType)
-  const dropdownActions = secondaryActions.slice(2).map((action) =>
-    isActionType(action)
-      ? {
-          ...action,
-          critical: action.critical || false,
-        }
-      : action
-  ) as DropdownItem[]
+  const visibleSecondaryActions = secondaryActions.slice(0, 2)
+  const dropdownActions = secondaryActions.slice(2).map((action) => ({
+    ...action,
+    critical: action.critical || false,
+  }))
+
+  /**
+   * Normalize the primary actions to be a list of groups
+   */
+  const primaryActions = useMemo(
+    () => normalizeItems(props.primaryActions ?? []),
+    [props.primaryActions]
+  )
+
+  /**
+   * Transforms the normalized primary actions into a format suitable for dropdown components.
+   * Each action group and its items are mapped to the expected dropdown item structure.
+   */
+  const primaryActionsDropdownItems = useMemo(() => {
+    {
+      return primaryActions.map(
+        (group): ActionBarGroup => ({
+          ...group,
+          items: group.items.map((item) => ({
+            value: item.label,
+            label: item.label,
+            icon: item.icon,
+            critical: item.critical,
+            description: item.description,
+            disabled: item.disabled,
+          })),
+        })
+      )
+    }
+  }, [primaryActions])
+
+  const singlePrimaryAction = useMemo(() => {
+    return primaryActions.length === 1 && primaryActions[0].items.length === 1
+      ? primaryActions[0].items[0]
+      : null
+  }, [primaryActions])
+
+  const getActionByValue = useCallback(
+    (value: string) => {
+      return primaryActions
+        .flatMap((group) => group.items)
+        .find((item) => item.label === value)
+    },
+    [primaryActions]
+  )
 
   return (
     <AnimatePresence>
@@ -139,49 +206,34 @@ export const OneActionBar = ({
               {warningMessage ? (
                 <Alert message={warningMessage} />
               ) : (
-                <>
+                <Fragment key="mobile-actions">
                   <MobileDropdown items={secondaryActions} />
-                  {primaryActions.length > 1 ? (
+                  {!singlePrimaryAction ? (
                     <F0ButtonDropdown
-                      items={primaryActions.map((action) =>
-                        isActionType(action)
-                          ? {
-                              value: action.label,
-                              label: action.label,
-                              icon: action.icon,
-                              critical: action.critical,
-                              description: action.description,
-                            }
-                          : { type: "separator" as const }
-                      )}
+                      items={primaryActionsDropdownItems}
                       onClick={(value) => {
-                        const action = primaryActions.find(
-                          (a) => isActionType(a) && a.label === value
-                        )
-                        if (action && isActionType(action)) {
-                          action.onClick?.()
-                        }
+                        const action = getActionByValue(value)
+                        ;(action as ActionType)?.onClick?.()
                       }}
                       size="lg"
                     />
-                  ) : primaryActions.length === 1 &&
-                    isActionType(primaryActions[0]) ? (
+                  ) : (
                     <F0Button
-                      label={primaryActions[0].label}
-                      icon={primaryActions[0].icon}
-                      onClick={primaryActions[0].onClick}
-                      disabled={primaryActions[0].disabled}
+                      label={singlePrimaryAction.label}
+                      icon={singlePrimaryAction.icon}
+                      onClick={singlePrimaryAction.onClick}
+                      disabled={singlePrimaryAction.disabled}
                       size="lg"
                     />
-                  ) : null}
-                </>
+                  )}
+                </Fragment>
               )}
             </div>
             <div className="hidden items-center gap-2 sm:flex">
               {warningMessage ? (
                 <Alert message={warningMessage} />
               ) : (
-                <>
+                <Fragment key="desktop-actions">
                   {dropdownActions.length > 0 && (
                     <Dropdown items={dropdownActions} />
                   )}
@@ -198,38 +250,25 @@ export const OneActionBar = ({
                         disabled={action.disabled}
                       />
                     ))}
-                  {primaryActions.length > 1 ? (
-                    <F0ButtonDropdown
-                      items={primaryActions.map((action) =>
-                        isActionType(action)
-                          ? {
-                              value: action.label,
-                              label: action.label,
-                              icon: action.icon,
-                              critical: action.critical,
-                              description: action.description,
-                            }
-                          : { type: "separator" as const }
-                      )}
-                      onClick={(value) => {
-                        const action = primaryActions.find(
-                          (a) => isActionType(a) && a.label === value
-                        )
-                        if (action && isActionType(action)) {
-                          action.onClick?.()
-                        }
-                      }}
-                    />
-                  ) : primaryActions.length === 1 &&
-                    isActionType(primaryActions[0]) ? (
+                  {!singlePrimaryAction ? (
+                    <>
+                      <F0ButtonDropdown
+                        items={primaryActionsDropdownItems}
+                        onClick={(value) => {
+                          const action = getActionByValue(value)
+                          ;(action as ActionType)?.onClick?.()
+                        }}
+                      />
+                    </>
+                  ) : (
                     <F0Button
-                      label={primaryActions[0].label}
-                      icon={primaryActions[0].icon}
-                      onClick={primaryActions[0].onClick}
-                      disabled={primaryActions[0].disabled}
+                      label={singlePrimaryAction.label}
+                      icon={singlePrimaryAction.icon}
+                      onClick={singlePrimaryAction.onClick}
+                      disabled={singlePrimaryAction.disabled}
                     />
-                  ) : null}
-                </>
+                  )}
+                </Fragment>
               )}
             </div>
           </div>