+++ packages/react/src/global.types.ts
@@ -1,0 +1,5 @@
+export type UrlString = `http://${string}` | `https://${string}`
+
+export interface DataAttributes {
+  [key: `data-${string}`]: string | undefined
+}