+++ packages/react/src/ui/InputField/InputField.tsx
@@ -138,6 +138,7 @@ const inputFieldStatusVariants = cva({
 })
 
 export type InputFieldProps<T> = {
+  autoFocus?: boolean
   label: string
   placeholder?: string
   labelIcon?: IconType